requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-barcode==0.15.1
qrcode==7.4.2
Pillow==10.0.0
python-dateutil==2.8.2
pytest==7.4.0
gunicorn==21.2.0
email-validator==2.0.0
Flask-Mail==0.9.1
